refactor: fix directory name
