+++ lib/provider/account.dart
@@ -1,5 +1,5 @@
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:person_note/model/auth/account.dart';
+import 'package:person_note/model/account/account.dart';
 import 'package:person_note/provider/auth.dart';
 
 final accountProvider = Provider<Account?>((ref) {
